Added Disc(InputAudioFile &audio) and Disc(CueDisc &cue) constructors remove loadFromCue method

+++ tests/test_inputaudiofile.cpp
@@ -40,7 +40,7 @@ void TestFlacon::testInputAudioFile()
     QFETCH(QString, duration);
     QFETCH(QString, format);
 
-    uint dur = duration.toInt();
+    uint dur = uint(duration.toInt());
 
     try {
         InputAudioFile ia(fileName);
